Add save of current workout plan
Improvements to home page. Rename repository, database. Change database Ids to long. Add autofocus of textfield in add ... dialogue

+++ app/src/main/java/com/anexus/perfectgymcoach/data/exercise/Exercise.kt
@@ -8,7 +8,7 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 @Entity
 data class Exercise(
-    @PrimaryKey(autoGenerate = true) val exerciseId: Int = 0,
+    @PrimaryKey(autoGenerate = true) val exerciseId: Long = 0L,
     val name: String,
     val primaryMuscle: Muscle,
 //    val secondaryMuscles: List<Muscle>,
